+++ docs/assets/UluFormActions.stories-xosirVXg.js
@@ -1,0 +1,55 @@
+import{_ as o}from"./UluFormActions-D5G1Jkah.js";import{U as r}from"./iframe-CfzwUndu.js";import"./preload-helper-BJwshlQW.js";const a={component:o,tags:["autodocs"]},t={render:e=>({components:{UluFormActions:o,UluButton:r},setup(){return{args:e}},template:`
+      <div class="form-theme">
+        <UluFormActions>
+          <UluButton primary text="Submit" />
+          <UluButton text="Cancel" />
+        </UluFormActions>
+      </div>
+    `})},n={render:e=>({components:{UluFormActions:o,UluButton:r},setup(){return{args:e}},template:`
+      <div class="form-theme form-theme--actions-right">
+        <UluFormActions>
+          <UluButton primary text="Submit" />
+          <UluButton text="Cancel" />
+        </UluFormActions>
+      </div>
+    `})};t.parameters={...t.parameters,docs:{...t.parameters?.docs,source:{originalSource:`{
+  render: args => ({
+    components: {
+      UluFormActions,
+      UluButton
+    },
+    setup() {
+      return {
+        args
+      };
+    },
+    template: \`
+      <div class="form-theme">
+        <UluFormActions>
+          <UluButton primary text="Submit" />
+          <UluButton text="Cancel" />
+        </UluFormActions>
+      </div>
+    \`
+  })
+}`,...t.parameters?.docs?.source}}};n.parameters={...n.parameters,docs:{...n.parameters?.docs,source:{originalSource:`{
+  render: args => ({
+    components: {
+      UluFormActions,
+      UluButton
+    },
+    setup() {
+      return {
+        args
+      };
+    },
+    template: \`
+      <div class="form-theme form-theme--actions-right">
+        <UluFormActions>
+          <UluButton primary text="Submit" />
+          <UluButton text="Cancel" />
+        </UluFormActions>
+      </div>
+    \`
+  })
+}`,...n.parameters?.docs?.source}}};const c=["Default","RightAligned"];export{t as Default,n as RightAligned,c as __namedExportsOrder,a as default};