+++ docs/assets/UluFormRequiredChar.stories-BLpzEeEh.js
@@ -1,0 +1,27 @@
+import{N as r}from"./iframe-CfzwUndu.js";import"./preload-helper-BJwshlQW.js";const s={component:r,tags:["autodocs"]},e={render:a=>({components:{UluFormRequiredChar:r},setup(){return{args:a}},template:`
+      <div class="form-theme">
+        <label>
+          My Label
+          <UluFormRequiredChar />
+        </label>
+      </div>
+    `})};e.parameters={...e.parameters,docs:{...e.parameters?.docs,source:{originalSource:`{
+  render: args => ({
+    components: {
+      UluFormRequiredChar
+    },
+    setup() {
+      return {
+        args
+      };
+    },
+    template: \`
+      <div class="form-theme">
+        <label>
+          My Label
+          <UluFormRequiredChar />
+        </label>
+      </div>
+    \`
+  })
+}`,...e.parameters?.docs?.source}}};const t=["Default"];export{e as Default,t as __namedExportsOrder,s as default};